Redo grid generation as a map
After talk with more experienced developers at the Elixir user group the
design of the grid has been altered. A map is used as the container to
allow more efficient use of random access algorithms. A tuple of X and Y
coordinates serves as the key to a cell which points to a list that can
contain links to other neighboring cells. Links could be represented as
atoms e.g. `:north, :south, :east, :west`.

Consider making a `Point` or `Coordinate` struct to enforce the
convention of X and Y coordinates matching columns and rows
respectively. Using these may help to reduce errors caused my reversing
the two values.